modules/kvs; Refactor kvs initialization of root
In core KVS file, instead of initializing root whenever the module
is loaded, instead initialize root whenever the first kvs get, watch,
sync, or fence request is made.

In other callbacks, adjust appropriately.  In some functions, such
as kvs.unwatch, do nothing if the root is initialized.  In functions
such as dropcache, nothing changes b/c it is not relevant if the root
is initialized or not.  In stats.get, return all zeroes to user.  In a
getroot request, it must initialize itself before returning the root
to the requestor.  The getroot request remains synchronous and recurses
up to the root of the TBON to initialize.