kvs: Add namespace parameters in rpcs
In KVS request rpcs, add namespace parameter that specifies
which namespace the RPC belongs to.  For the time being,
it is always the primary namespace.

In KVS server, adjust code that previously hardcoded the
primary namespace to instead use the namespace provided in
the rpc.

Rename KVS_PRIMARY_ROOT_KEY to KVS_PRIMARY_NAMESPACE as consequence.