modules/kvs: Refactor event names and subscription
Modify kvs.setroot and kvs.error events to be specific to each
namespace, so that nodes without the namespace don't receive
the events.

Update event subscription to only subscribe to the specific events
meaningful to the module.

In addition, don't subscribe to any events until the first root
is initialized.