jobspec: replace minus operator with multiplication
Make the implementation come to compliance with RFC14's Content
Rule section. It currently does not have the subtraction operator
listed, but multiplication:

===
Content Rules

$complex_range = {
    "min" : 1..,
    "max" : 1..,
    "operator" : ( :"+" | :"*" | : "^" ),
    "operand" : 1..,
}
===

The role of an operator is:
An operator applied between min and max which returns the next
acceptable value.
It seems using an operator that can return a decreased
or the same next value doesn't make whole lot of sense to me.

Tested with resource-query and this one-liner seems enough.