t/t5000-valgrind.t: run broker with "libtool e"
Problem: valgrind script fails due to missing lt-flux-broker
on libtool 2.4.6 i686.

The lt- prefix on .libs executables is missing on this platform.
libtool docs say it may or may not be present.

Use libtool --mode=execute (abbreviated as "libtool e") to get
the same effect portably, as described in the Libtool manual,
Section 3.4: Debugging Executables.