modules/kvs: Refactor lookup_create()
Remove root_dir parameter to lookup_create().  It was only used
for one corner case, which in reality is not needed.

Instead, when user passes in root dir path (i.e. "."), it will return
the root reference that was passed in by the user.  It does not need
to return the "real" root reference.

Update unit tests appropriately.

Fixes #1297