modules/kvs: Add namespace to internal commit API
Refactor commit_mgr_create() to pass in the name of the namespace
for the commit mgr.  While not used internally, it is critical to
know the namespace a commit is attached to for other purposes, most
notably on replays.

Add get function commit_get_namespace() to get a namespace.

Add appropriate unit tests.

Adjust callers appropriately.