modules/kvs: Refactor kvsroot with remove flag
In preparation for kvs remove namespace support, place remove flag
in kvsroot to indicate if kvsroot is marked for removal.

Add new function lookup_root_safe() which looks up root and checks
if a root is marked for removal.  For the time being, all code
assumes a root cannot be marked for removal.