modules/kvs: Code re-org in core kvs file
Recent changes to the core KVS file lead to a change in how
many of the functions in the file were called and/or used.  Those
functions now seem "out of place" given the current code.  Move
functions around and re-organize in general.

No code has changed in this commit.  Functions have only been
moved around.