broker/boot_config: add config file boot method
If a "boot.method" attribute is set to "config", bootstrap
overlay parameters from a config file instead of PMI.

If it is set to PMI or is unset, the previous PMI boot
method is executed.

If config is selected, the attribute "boot.config_file"
attribute must point to a TOML configuration file.
See boot_config.h.