broker: create broker.rundir early
Problem: create_rundir() is called both within boot_pmi(),
and after bootstrap.

The call within boot_pmi() was added so that the session-id
obtained from PMI could be used in the directory name,
and the rundir could be used in ipc:// socket paths.
However the session-id is no longer used in the path, so we
can make single call before bootstrap.