t/t1105-proxy.t: address sensitivity to test env
Problem: four subtests of t1105-proxy.t fail when
run on a system that is also running a Flux system
instance.

These tests use the construct
  FLUX_URI=value VAR=value ... test_must_fail command

It looks like the test_must_fail shell function
actually does not transmit the environment variable
settings to "command", and since FLUX_URI is unset
by the test script, the command attempts to connect
to the system instance and succeeds.

This example, which prints nothing and exits with a
code of 1, demonstrates the problem with the construct:

 #!/bin/sh
 foo() {
     "$@"
 }
 X=1 foo printenv X

Solution: run the command directly and test its exit
code, instead of running it with test_must_fail().