modules/kvs: Refactor fence names array
With recent changes, fence no longer needs to store an array
of names, it need only store the single name of the fence.

As a consequence, fence_get_json_names() is now fence_get_name()
and returns a const char * instead of a json_t *.

With this change, the fence data structure now more closely maps
to what a "fence" is, as it can only contain one name, and the
set of operations on that name.

Update unit tests appropriately.