modules/kvs: Rename "commit" data structure
For consistentcy with other portions of code in Flux,
rename the internal "commit" API to avoid confusion with an
actual "commit" RPC.  Rename it to the "kvstxn" API.

No logical code changes occurred in this commit.  Search and
replacing "commit" to "kvstxn" was mostly done.  Some variables
were renamed as a result (e.g. "cm" with "ktm").  Some
functions were also renamed as a result for clarity (e.g. instead
of "txn", used "transaction" in some locations).

Minor comment adjustments as well.