modules/kvs: Preserve orig kvstxns during merge
Do not destroy transactions after they have been merged.  Instead
flag them as components of a larger merge.  When the kvstxn
of a set of merged transactions completes/is removed, at that point
in time remove all of the components of the larger merge.

As a consequence of this change and for optimization purposes, once
a merger of transactions has occured, there can no longer be any more
mergers until the head merged transaction has completed.  If this were
not done, the ready queue would constantly be iterted through and
new head merged transactions would be created.  This can be optimized
at a later time.

Add unit tests.