modules/kvs: try orig transactions on kvstxn error
In kvstxn_mgr_remove_transaction() support flag for user to fallback
a merged kvstxn to the original transactions that made up the merge.

By doing so, the user need not send an error to all transactions merged
into that kvstxn.  Instead, each of the original transactions
can be replayed individually, and an error will only be sent to
the offending commit/fence transaction.

Support kvstxn_fallback_mergeable() so user knows if a kvstxn can
be falled back on.

In kvstxn_apply(), take advantage of this by not sending an error
when a kvstxn's merging can be falled back on.  As an exception,
do not fallback if it's a "death"-like error (e.g. ENOMEM).

Add internal kvstxn API unit tests.

Fixes #1337