modules/kvs: Lookup root ref via namespace
In lookup_create(), instead of always requiring the user to
specify the root reference, instead make it optional.  If the
user does not specify it, find it via the namespace.

In this patch, it is assumed the namespace specified by the
user always exists in the kvsroot_mgr_t.

Update unit tests accordingly.