build: add --disable-jobspec opt-out configuration
Problem: flux-core silently skips building its jobspec
library if libyaml-cpp is not found, but flux-sched
requires the jobspec library in order to build,
and when it is missing, it's not obvious what project is
supposed to provide it.

Add configure --disable-jobspec option to opt out of
building the jobspec library.  Unless opted out, missing
jobspec prereqs are fatal.

Fixes #1360