cmd/flux-submit: indicate to sched when -N not set
When only -n is used with flux-submit, the wreck implementation
sets a default value for nnodes to the minimum of the number
of tasks requested and the total number of nodes in the session.
This unfortunately loses the context that the user didn't explicitly
request a number of nodes.

For flux-submit only, change this behavior and set the nnodes value
in the request to 0 if -N, --nnodes was not set by the user.

Fixes #1368