wreck: save -n,-N,-t options in kvs for submission
For posterity and possibly scheduler use, save the raw value
of the -N, --nodes; -n, --ntasks, and -t, --tasks-per-node options
used in flux-submit and flux-wreckrun to the LWJ kvs directory
under an `opts.` key. If an option wasn't used at all, then
there will no corresponding entry in opts. for that option.