libjsc/wreck: add internal wreck class
1) wreck_set_notify()

Provide a mechanism for registering a callback that
will be called for selected wreck job state transitions.
A job record is created in memory on receipt of the first
monitored state transition, and persists until a terminating
state transition occurs.

Upon creation of the job record, the requested ntasks,
nnodes, and walltime is fetched from the KVS.  This is hidden
from the caller and performed asynchronously, with the
initial state callback made after KVS fetches have completed.

2) wreck_job_aux_set(), wreck_job_aux_get()

The user may attach 'aux' data to a job record, with optional
destructor.

3) wreck_job_incref(), wreck_job_decref(), wreck_job_lookup()

Provide a mechanism for job records to be looked up in the
internal hash, and to be allowed to persist beyond the
occurrence of a terminating state.

4) wreck_launch()

Add a function to assign resources to a wreck job in the KVS
and tell wreck to start it.

Resources are represented as a JSON string containing an
array of objects indicating the assigned ranks and number
of cores on each, e.g.
  [{"rank":0,"corecount":4},{"rank":1, "corecount":4"}]
This corresponds directly to wreck's expected resource
assignment in the KVS:
  lwj.<id>.rank.0.cores = 4
  lwj.<id>.rank.1.cores = 4

After the KVS update completes, a "wrexec.run.<id>" event
is generated, to instruct wreck to execute the job on
the specified ranks/cores.  The function returns immediately,
and the event is sent once a response is received for the
KVS commit.

For now, export wreck_* from libflux-core.so so that this
can be used directly.  Wreck is going away so this is temporary.

5) wreck_set_state()

wreck_set_state() updates the KVS job schema's state
variable, then publishes a wreck.state.<state> event.

The function returns immediately once the KVS commit
request has been sent.  A continuation asynchronously
handles the commit response, and sends the event.