wreck: skip reserved state for submitted job
The 'reserved' state is meant only for a reserved KVS directory
for a job which has not yet been submitted or run (i.e. reserved
for wreck as writer). In the case of jobs submitted via flux-submit
this state is unecessary, so remove the initial reserved state for
submitted jobs, and the corresponding duplicated code that was
a result.