modules/kvs: Add root_ref to internal walk levels
In internal walk_level_t in the lookup API, add the root reference
and the root_dirent of the walk level to each walk_level_t.

As a consequence, the "root_dirent" in the lookup_t is no
longer needed.

This refactor allows each walk level to eventually have its own
independent root reference, which could be different depending
on the namespace.