wreck: switch to rcalc for count of tasks to run
Problem: wrexecd assumes that the rank.N.cores count assigned by
the scheduler is the same as the number of local tasks to run
on the local rank.

Solution: Switch wrexecd to using the `rcalc` code to distribute
the total ntasks requested for the job across the global resources
assigned. The code first checks for an "R_lite" key in the Rlite
format, and if that is not avilable, falls back to the rank.N.cores.

Fixes #1378