wreck: don't ignore lwj.environ in wreckrun/submit
wreck execution first reads a global lwj.environ table and uses
that as the default environment for all jobs, overridden by any
environment variables encoded in the per-job `environ` table.

Currently, flux-wreckrun and flux-submit do not check for lwj.environ
and instead export the entire (filtered) current environment to each
job, so the lwj.environ is largely going ignored.

Fix this by first reading a "default environment" from lwj.environ
in wreckrun/submit, and only push variables that are not already
the same as the default env to the job. This should cut down on
KVS size for large numbers of jobs with largely the same, but slightly
different, environments. (If all environment tables are the same, then
this change does nothing since all environ entries in KVS will be
squashed by CAS)