wreck: Add -S, --skip-env option to wreckrun/submit
Add an option -S, --skip-env to submit and wreckrun to skip the
export of current environment to the job. This can be used to speed
up submit/wreckrun when the global lwj.environ is sufficient, since
that default table doesn't need to be fetched from the KVS before
submitting the job to the instance.