libkz: defer errors from continuations
Problem: errors that occur in context of a KVS continuation
in the read path were not saved so they could be returned to
the user on the next kz_get(), kz_get_json(), or kz_close() call.

Check for saved error in kz_get(), kz_get_json(), and kz_close().
When a continuation encounters an error, save it, and call the
user's callback.

A couple of inappropriate log messages were removed also
and some error paths further cleaned up.