lua/t1003-iowatcher.t: fix test error check
Problem: lua iowatcher test 5 'iowatcher returns error
correctly' fails after kvs_watch deferral change.

This test attempts to create an iowatcher on a non-directory
and expects iowatcher creation to fail.  Now that the
the lookup is no longer performed in the context of
iowatcher creation, creation succeeds.

Change the test's iowatcher callback to stop the reactor
if 'err' is non-nil.  Then change he test to run the reactor
and expect an ENOTDIR failure.