test/kz: [cleanup] rename kzutil to kzcopy
Problem: kzutil --attach is not used by any tests.

Drop the attach code which is hard wired for wreck's
stdio naming convention, and is not used by any test.
This leaves kzutil --copy as the main mode, so drop
the --copy option and rename the utility to kzcopy.

Update users.