cmd/flux-wreck: add "uri" subcommand
Add a subcommand that lists the lwj.X.Y.Z.flux.remote_uri
value, if available.

Usage is similar to flux wreck ls, e.g.
  flux wreck uri [-n, --max=count] [b, --bare] [JOBIDS...]

If called with --bare, only the URI for exactly one job is
listed, by itself for easy parsing.  It is an error if there
is not exactly one job specified, or if the job is not a
Flux instance.

If called without --bare, each job is listed with minimal
state information.  The FLUX_URI fields is left blank for
jobs that are not Flux instances.

$ flux wreck uri
ID NTASKS STATE     FLUX_URI                                 COMMAND
 1      1 exited                                             hostname
 2      1 exited                                             hostname
 3      1 exited    ssh://jimbo//tmp/flux-qCbPz5             flux