wreck/io.lua: capture and log errors in iowatcher callback
Capture the error string parameter if passed to the iowatcher callback
in the wreck/io.lua implementation. If an error is passed to the handler,
then log it and disable that iowatcher (i.e. consider it closed).

The error is logged to stderr unless a new "log_err" function is passed
through the constructor, in which case errors are logged via this
function.

Fixes #1427