libjsc: add jsc_query_rdesc_efficiently()
Problem: JSON encode/decode of JSC rdesc objects proved
to be a significant bottleneck affecting scheduling
performance for high throughput workloads.

Add jsc_query_rdesc_efficiently() to allow nnodes, ncores,
ntasks, and walltime to be queries directly.

This was solved in a different way in current flux-sched
master, but this is needed as a temporary workaround until
older forks of flux-sched used with the Splash application
catch up with master per @trws.

Fixes #1455