modules/job: exec wrexecd with --kvs-guest option
Arrange for the FLUX_JOB_KVSPATH environment
variable to point to the guest KVS namespace
in a launched program, rather than the KVS path
in the primary namespace.

This changes the KVS path used to look up a
sub-instance's FLUX_URI from
  <job_kvs_dir>.flux.remote_uri
to
  <job_kvs_dir>.guest.flux.remote_uri

Update flux-wreck and t2003-recurse.t accordingly.