t/t2000-wreck: fix race condition in attach -n test
Problem: test for `flux-wreck attach --no-follow` has a race
condition, since there is no synchronization between when task
output makes it to kvs, and the run of `attach --no-follow`.
We still want to ensure that attach can print output that has
already made it to the kvs, so we don't want to accept only
partial output as a sign of success.

Fix: retry the `attach --no-follow` in a loop until all data has
made it into kvs.

Fixes #1459