sharness: Always generate broker.log in test_under_flux
Problem: The test_under_flux() sharness function fails to create a
broker.log for the test when the --verbose flag is used. This is
counterintuitive.

I think this was meant to send dmesg log messages to stderr when
--verbose was used. However, this no longer is the case. Additionally,
non-verbose mode used the broker's `-q, --quiet` flag, which apparently
does nothing.

Instead of the current confusing operation, just remove the use of -q,
and always write a broker.log file for tests using test_under_flux
with log-forward-level=7.