sharness: don't remove broker.log with --debug
When using test_under_flux a broker.log file is created to capture
all log messages into a log file. However, there is no way to preserve
this file for successful tests, as other test products can be
preserved with --debug.

Set the broker.log for tests to only be added to cleanup when
--debug has not been used.

Fixes #1288