wreck: add wreck_job class
Add a 'struct wreck_job' data structure, useful for
packaging up job information that needs to be passed
through the void * argument of a continuation.

The class includes functions for keeping wreck_job's
in a zhash_t, should that be needed later on for
tracking active jobs.

An "aux" accessor allows ancillary data to be attached
to the wreck_job, with destructor.