broker: add synchronous event RPC
Problem:  when multiple module instances publish
events, with the order coordinated by a third party,
there is no way to ensure that the events sent
sequentially in time processed in the same order when
they are rebroadcast.

Add a rank 0 broker service to publish events with
a response indicating that the event has received
a sequence number.

The "fire and forget" event transport is still the
primary one.  This is here primary to support the
wreck job module.  If it turns out to be more useful,
we can add it to the Flux API.

Fixes #342.