wreck: track/list state of active jobs
Track the state of active jobs in a hash in the job
module.  A job is no longer active when it reaches
the "complete" or "failed" states.

Add a 'job.list' RPC which returns a list of active
jobs in the form returned by wreck_jobs_list(),
passing max, include, and exclude parameters in
the request to the function.