wreck: add wreck_job_list() to wreck_job class
Add a function to produce an object containing
a JSON array that is a sorted list of jobs
from a zhash_t of struct wreck_job entries.

Each time the state of a job is updated with
wreck_job_set_state(), set "mtime" of the job
to the current time.  Then sort the jobs in
the array in mtime order.

Limit the number of jobs returned in array
to 'max' if > 0.

If non-NULL, 'include_states' and 'exclude_states'
are parsed as comma-separated lists of job states
to include or exclude from the list, with
blacklist/whitelist semantics.