wreck/lua: use wreck.id_to_path instead of rpc in io.lua
Problem: wreck/io.lua uses a hand-coded rpc to the job.kvspath service
to lookup the kvs path for a jobid if needed, resulting in multiple
call sites to check for fixes and proper args to the rpc.

Use the wreck.id_to_path() function built-in to wreck.lua instead. This
not only consolidates the rpc to a single call site, but also makes
use of memoization of the wreck version of this function.