wreck: use cmb.exec not fork/exec in job module to spawn wrexecd
Problem: The job module spawns wrexecd on ranks involved in a job by
direct fork/exec. Not only does this duplicate code from the cmb.exec
service, but this is a fork from a threaded program, it may be
problematic to reap exit status of wrexecd since the job module
may not be successful in registering child watchers.

Switch from direct fork/exec in job module to and rpc to the `cmb.exec`
service on the local broker. Use the new flux_future_reset() facility
to accept multiple responses from the `cmb.exec` service. Finally
destroying the job and the future after the remote wrexecd has exited.

For now non-zero exit codes and failure of exec(2) for wrexecd are
simply logged, and no further action is taken.