wreck: fix event payload for job.submit-nocreate
Problem: The event generated as a result of `job.submit-nocreate`
rpc from flux-wreckrun had ncores and ngpus set to 0 since wreckrun
did not forward these values along in the message of the rpc. This
results in confusion for sched, the main use case for the
submit-nocreate service.

Since the wreck/job module now can cache active jobs, there is no
longer a reason to require the caller to forward along all data in
the job.submit-nocreate rpc. Instead, have the job.create rpc
preemptively add the created struct wreck_job to the active_jobs
hash, and have the `job.submit-nocreate` callback fetch the fully
instantiated job by jobid from the hash.

This assumes that job.submit-nocreate will be called on the same
rank as job.create, but for the wreckrun case this is certainly
true.

Fixes #1491