libflux/future: add flux_future_reset()
Problem: there is no way for a future to be "unfulfilled"
in cases where there are multiple events to be handled by
the future.

Add flux_future_reset(), which sets up the future to be
fulfilled again, and invalidates any result currently stored
in the future container.

Allow flux_future_then() to be optionally called again
after flux_future_reset(), to set a timeout or change
the continuation callback.