libflux/message: add checks to flux_msg_set_payload
Problem: flux_msg_set_payload() allows a FLUX_MSGFLAG_JSON
payload to be added that is not valid JSON.

Move the cursory check (not a full JSON object parse) from
flux_msg_set_json() to flux_msg_set_payload(), conditional
upon FLUX_MSGFLAG_JSON being set.

In addtion, veryfiy that the '\0' string terminator is
included in the payload, and that the string ends with '}'.

Finally, add an EINVAL check that the msg parameter != NULL,
since that would cause a segfault.