cmd/flux: reverse logic of in-tree detection
Problem: when flux is launched by spindle, /proc/self/exe
no longer refers to $prefix/bindir and flux attempts
to use paths relative to its build directory to access
the rest of flux.

Instead of comparing /proc/self/exe's symlink target
to $prefix/bindir, compare to ${abs_top_builddir}/src/cmd.
That way the installed executable can be relocated
without harm, and the "special" case is the one where
flux is run out of the build tree, which seems appropriate.

The problem launching flux with spindle is discussed
in issue #1514.