libjsc: rename J_STOPPED to J_SYNC
Problem: The libjsc J_STOPPED state ("stopped") is not actually set
by the wreck execution system. It is presumed this was meant to be
the "sync" state set by wreck when tasks are all running but stopped
in exec(2).

Update JSC to use "sync" and J_SYNC instead of "stopped"/J_STOPPED.