libjsc: pull in shortjansson.h from sched
To prepare for conversion of libjsc from json-c
to jansson, pull in shortjansson.h wrappers
from flux-sched project and adapt to local
environment.

This allows libjsc to be converted in several
steps;  otherwise, it's got to be done in one
giant patch to avoid breaking git bisect.

Drop Jtostr() - we will convert those directly
to json_dumps () since a free() is required now.