libjsc: consult active_jobs hash in jobid_exist()
Problem: every jsc query first calls jobid_exist(),
which performs a synchronous KVS lookup on the job's
KVS directory.  This is unnecessary if the job
is already in the active_jobs hash.

If job is in the active_jobs hash, return success
immediately.

Also, if the KVS lookup is performed, call
flux_future_get() rather than flux_kvs_lookup_get_dir()
since the latter decodes the directory, and we are
only going to throw it away.