wreck: store options as single json object
Problem: The wreck job "options" are stored in the KVS as separate
keys, and their values are fetched back-to-back in a complicated
function in wrexecd.

Store the options more simply as a single JSON object so that it
can be fetched in a single KVS operation, and the gathering of
options simplified.