modules/job-ingest: add module for job ingest
Add a module that handles job-ingest.add RPCs to
add new jobs to the KVS using a rudimentary form of
the RFC 16 job schema foramt.  The user is returned
a jobid based on the FLUID proposal, which allows
64-bit id generation to occur in parallel across ranks,
while retaining a loose ordering of ids based in the
time submitted.  KVS per-job directories are generated
using the DOTHEX FLUID encoding, e.g.

job.active.0000.011c.ae00.0002

The instance owner and any user with ROLE_USER may
submit jobs.  Jobs must be signed, and the user
authenticated as the submitter (by the connector) must
match the signature, but the job signature is not
authenticated at ingest time.  The connector-authenticated
userid is recored in the KVS under the "userid" key.
The signed blob is recorded under the "J-signed" key.

The job submission consists of signed RFC 14 jobspec,
which is validated by temporarily instantiating a C++
Jobspec object and recording any parse errors.
The parsed Jobspec may be further validated in the
future, for example to find resource requests that can
not be fulfilled, but for now we ingest all valid jobspec.
(Validation is performed in a standalone .cpp file linked
against libjobspec.la.  The standalone C++, which exports
a validate function callable from C, is compiled with the
C++ compiler; automake then knows to link job-ingest
against libstdc++).  The unwrapped jobspec is written
to the KVS under the "jobspec" key.

The module is completely event driven, and KVS overhead
is reduced and ingest rate increased by batching job-ingest.submit
requests that arrive toether within 10ms.

A "job-ingest.submit" event is generated after the KVS commit
which contains an array of new jobids.  This can be consumed
by the job-manager module in the future, which will handle
listing jobs for users, and informing the scheduler when
new active jobs have been submitted.