cmd/flux-aggregate: allow many types of value arguments
Loosen restrictions on the types of values aggregated by the
flux-aggregate test utility.

Allow value argument on flux-aggregate cmdline to be a string or a
numeric value.

Add a `-e, --execute=CODE` argument to allow values that are created
from arbitrary Lua CODE. This allows generation of arrays and complex
objects for aggregator testing.