libutil/mnemonic: add Orin Tirosch's mnemonic encoder
Mnemonic encoder allows 32 bits of data to be converted into
three words from a ~1630 word vocabulary, with words chosen
to be easy to understand over the phone and recognizable
internationally.

From original project here (now defunct):

http://web.archive.org/web/20101031205747/http://www.tothink.com/mnemonic/

This will allow 64 bit FLUIDs to be represented as a 6 word
phrase, which may them easier to communicate (especially thinking
of incident reports for problems on classified computer networks).

Originally proposed by @grondo and discussed in #470.