libflux/message: rename set/get_json to _string
Problem: it's not convenient to send a C string
message payload.

Current code that accepts a 'json_str' does a cursory
check that the string looks vaguely like an
encoded JSON object.  Better to just drop this check
and allow any type of string in the payload.

Rename and drop the cursory check:
  flux_msg_set_json() -> flux_msg_set_string()
  flux_msg_get_json() -> flux_msg_get_string()

N.B. it's convenient to include the terminating \0
in the payload so accessor functions can return the
string as a (const char *) without copying it first
to add the \0 character.

Update users and tests.