libutil/fluid: detect bad mnemonic input
Problem: fluid_decode (type=MNEMONIC) does not fail if
presented with words not in its dictionary or a phrase
too short to represent a uint64_t.

mn_decode()'s inline documentation is incorrect:

/* Return value:
 *  This function may return all the value returned by mn_decode_word_index
 *  plus the following result code:
 *
 * MN_EWORD  - Unrecognized word.
 */

It actually returns the number of bytes successfully
written to the output.

Require mn_decode() to return 8 in order for fluid_decode() to
be successful.