reactor: add flux_buffer_write_watcher_close
Add a close method to the flux_buffer_write_watcher interface. The
close method writes a pending EOF to the buffer watcher, which will
automatically close the underlying file descriptor once the buffer
is empty.

After the fd is closed, the user's callback is called with FLUX_POLLOUT
and a new accessor flux_buffer_write_watcher_is_closed () can be called
to retrieve the errno (if any) from the close(2) system call.

Fixes #1547