libflux: add implementation for composable futures
Problem: The Flux futures API provides no facility to assist in the
creation of composite futures (futures which are fulfilled as a result
of completion of other futures). This results in user frustration and
duplicated boilerplate code throughout the Flux codebase, as well as
presumably in eventual user code.

This small extension to the futures API extends the futures type with the
concept of "composite future" containers, which are automatically fulfilled
when either all futures in the container are fulfilled (when_all), or when
one of any futures in the container are fulfilled (when_any). Since these
containers are also of type flux_future_t, they can be further composed
to create a single synchronization handle for complex operations.

The interface adds two new future constructors for the two composite
container types

  flux_future_t *flux_future_wait_all_create (void);
  flux_future_t *flux_future_wait_any_create (void);

Along with standard composite object methods to push new futures into the
container by name:

 int flux_future_push (flux_future_t *f, const char *name, flux_future_t *child)

as well as an iterator interface and getter to traverse a composite future's
children:

 const char * flux_future_first_child (flux_future_t *cf);
 const char * flux_future_next_child (flux_future_t *cf);

 flux_future_t *flux_future_get_child (flux_future_t *cf, const char *name);

Higher level interfaces that return a composite future are expected to
provide a wrapper call for flux_future_get(), which will then traverse
the composite's children to build a possibly complex result.