libflux: add implementation for sequential composition of futures
Problem: The Flux futures API provides no facility to assist in
sequential composition of futures, or future chaining. Currently,
chains of futures are implemented simply by a recursive chain
of "then" callbacks. The drawbacks to this appraoch is the intent
of the entire chain isn't easily evident from the function that
creates the initial future, and there is no way to easily create
a synchronization handle (new future) for the final event in the
chain.

This extension to the base flux_future_t type provides two new
functions which allow sequential composition, or the creation
of chained futures:

 flux_future_t *flux_future_and_then (flux_future_t *f,
				      flux_continuation_f cb, void *arg);
 flux_future_t *flux_future_or_then (flux_future_t *f,
                                     flux_continuation_f cb, void *arg);

These functions return a chained future tied to their nominal
predecessor, `f`, which can then be "continued" from the provided
continuation function `cb`.

The `flux_future_continue(3)` interface is then used to "continue"
the chain at a later time within the continuation callbacks `cb`.

This not provides an easy method to obtain a flux_future_t on
which to synchronize after one or more steps in a chain, but results
in clearer code, since instead of spreading the logic through
multiple callbacks, the intent is clear at the initiating call-site,
e.g.

 flux_future_t *f = flux_rpc (...);
 flux_future_t *f2 = flux_future_and_then (f, step2, arg);
 flux_future_t *f3 = flux_future_and_then (f2, step3, arg);

 return (f3);

Now f3 can be fulfilled as a result of a future not created until the
function step3(), which has not executed yet.

The "or_then" function is provided for cases where future `f` is fulfilled
with an error and the implementor would like to attempt recovery or
handle the case in a method other than the default -- propagating the
error immediately to the next error in the chain.