libflux: defer errors from init in flux_future_then
Remove the special-case code in flux_future_then() which handles
futures that have been fulfilled with error by the time the initialization
callback is complete.

This defers the handling of errors set before or during the init callback
to the provided continuation callback, instead of returning the error
immediately from flux_future_then().

This changes what was presumed to be an arbitrary decision during early
futures development on the Flux API. In either case, the error will
be handled, however, handling all errors in the continuation callback
seems slightly easier than tracking errors on futures in two places.