modules/kvs-watch: new module for KVS watch service
Add a new module 'kvs-watch' to track KVS commit events
and notify interested parties when the KVS changes.

This was implemented separately from the main KVS module
to avoid adding more complexity.  Eventually this service
can replace the current KVS watch machinery, but for now
it is built up independently.

The initial version implements a "kvs-watch.getroot"
method that sends an immediate response containing the
current root, and additional responses each time the watched
namespace changes.  It interacts with the main kvs module by
sending it an initial "kvs.getroot" RPC, and then tracking
changes via published "kvs.setroot" events.  Watcher state
is shared among multiple clients.  Event subscriptions are
established on-demand.

The user may make a "kvs.getroot" RPC to receive the
current root information in one response, or a
"kvs-watch.getroot" RPC to receive the current and
future changes in multiple responses.  These services
use compatible protocols.  The API will select which
service to used based on a flags setting.  Eventually
this approach will allow the now separate flux_kvs_watch()
API to be folded in with flux_kvs_lookup().

The stream of responses continues until the request is
explicitly cancelled or the client disconnects.  Explicit
cancellation consists of sending a "kvs-watch.cancel"
request containing the requestor UUID and matchatg.  This
request does not receive a response;  the response stream
is terminated with an error response (errnum = ENODATA)
as specified in RFC 6.  The sender should wait for this
error response before recycling the original request matchtag.