libflux/rpc: add flux_rpc_get_matchtag()
Problem: It is difficult to implement multi-response
streaming RPC services that allow the client to terminate
the stream, because the flux_rpc() user (client) cannot
reconstruct the (sender-id, matchtag) tuple that uniquely
identifies a request on the server end.

Add an accessor for the flux_rpc() request matchtag.
The client can send a new request to the server that
contains this matchtag in its payload.  The server can
then extract the sender-id from the request route stack
and use the tuple to look up its state for the "stream"
and cancel it, sending an ENODATA response per RFC 6.

Note: this is not a basis for the "RPC cancellation"
feature that has been deferred for some time, since it
requires the client and server to implement the cancel
RPC out of band with respect to the original RPC.
However, practical experience with this may inform
a future design.